:sparkles: feat: Add particle effects on player death
In this commit, we've added particle effects to enhance the visual feedback when the player character dies in the game. A new particle emitter has been added to the `play.ts` file, which is triggered when the player dies. The emitter uses a new 'pixel' asset, which has been added to the `load.ts` file. The player character is now also destroyed upon death.